requirements.txt
azure-functions
beautifulsoup4
pandas
azure-storage-blob
azure-cosmosdb-table
azure-cosmos